refactor(control): split runtime supervision
